Refactored KeyStorage
