+++ Source/SecureChat/KeyStorage.swift
@@ -8,37 +8,16 @@
 
 import Foundation
 
-@objc(VSPKeyStorage) public class KeyStorage: NSObject, KeyStorageProtocol {
-    private let keychainKeyStorage: VSSKeyStorage
+@objc(VSPKeyStorage) public protocol KeyStorage: NSObjectProtocol {
+    func store(_ keyEntry: KeyEntry) throws
     
-    init(keychainKeyStorage: VSSKeyStorage) {
-        self.keychainKeyStorage = keychainKeyStorage
-        
-        super.init()
-    }
+    func update(_ keyEntry: KeyEntry) throws
     
-    public func store(_ keyEntry: KeyEntry) throws {
-        try self.keychainKeyStorage.store(VSSKeyEntry(name: keyEntry.name, value: keyEntry.value))
-    }
+    func loadKeyEntry(withName name: String) throws -> KeyEntry
     
-    public func update(_ keyEntry: KeyEntry) throws {
-        try self.keychainKeyStorage.update(VSSKeyEntry(name: keyEntry.name, value: keyEntry.value))
-    }
+    func existsKeyEntry(withName name: String) -> Bool
     
-    public func loadKeyEntry(withName name: String) throws -> KeyEntry {
-        let keychainEntry = try self.keychainKeyStorage.loadKeyEntry(withName: name)
-        return KeyEntry(name: keychainEntry.name, value: keychainEntry.value)
-    }
+    func deleteKeyEntry(withName name: String) throws
     
-    public func existsKeyEntry(withName name: String) -> Bool {
-        return self.keychainKeyStorage.existsKeyEntry(withName: name)
-    }
-    
-    public func deleteKeyEntry(withName name: String) throws {
-        try self.keychainKeyStorage.deleteKeyEntry(withName: name)
-    }
-    
-    public func getAllKeysTags() throws -> [Data] {
-        return try self.keychainKeyStorage.getAllKeysTags()
-    }
+    func getAllKeysTags() throws -> [Data]
 }
